♻️ Refactor files, add real and RMQ options
