+++ src/index.ts
@@ -1,4 +1,5 @@
-import { fuQuPubSub } from './lib/queue/pubsub';
-import { fuQuRabbit } from './lib/queue/rabbit';
+import { fuQuPubSub } from './lib/adapters/pubsub';
+import { fuQuRabbit } from './lib/adapters/rabbit';
+import { fuQuMemory } from './lib/adapters/memory';
 
-export { fuQuPubSub, fuQuRabbit }
+export { fuQuPubSub, fuQuRabbit, fuQuMemory }
